Calculate letters and calls
