+++ lib/laa_crime_forms_common/pricing/nsm/wrappers/letter_or_call.rb
@@ -1,0 +1,15 @@
+require_relative "base"
+
+module LaaCrimeFormsCommon
+  module Pricing
+    module Nsm
+      module Wrappers
+        class LetterOrCall < LaaCrimeFormsCommon::Pricing::Nsm::Wrappers::Base
+          wrap :type, Symbol
+          wrap :claimed_items, Integer
+          wrap :assessed_items, Integer
+        end
+      end
+    end
+  end
+end
